adapt to new externalized snapshot lib
