--- localstack/testing/snapshots/__init__.py
@@ -1,3 +1,0 @@
-from .prototype import SnapshotAssertionError, SnapshotMatchResult, SnapshotSession
-
-__all__ = [SnapshotSession, SnapshotMatchResult, SnapshotAssertionError]
