+++ tests/aws/services/iam/test_iam.snapshot.json
@@ -44,7 +44,7 @@
     "recorded-content": {
       "invalid-json": {
         "Error": {
-          "Code": "MalformedPolicyDocument",
+          "Code": "MalformedPolicyDocuments",
           "Message": "This policy contains invalid Json",
           "Type": "Sender"
         },
