+++ tests/conftest.py
@@ -4,9 +4,19 @@ from typing import TYPE_CHECKING
 import pytest
 from _pytest.config import PytestPluginManager
 from _pytest.config.argparsing import Parser
+from localstack_snapshot.pytest.snapshot import is_aws
+from localstack_snapshot.snapshots.transformer import RegexTransformer
+
+from localstack.constants import TEST_AWS_REGION_NAME
+from localstack.testing.snapshot.transformer_utility import (
+    SNAPSHOT_BASIC_TRANSFORMER,
+    SNAPSHOT_BASIC_TRANSFORMER_NEW,
+    TransformerUtility,
+)
+from localstack.utils.bootstrap import is_api_enabled
 
 if TYPE_CHECKING:
-    from localstack.testing.snapshots import SnapshotSession
+    from localstack_snapshot.snapshots import SnapshotSession
 
 os.environ["LOCALSTACK_INTERNAL_TEST_RUN"] = "1"
 
@@ -14,7 +24,7 @@ pytest_plugins = [
     "localstack.testing.pytest.cloudtrail_tracking",
     "localstack.testing.pytest.fixtures",
     "localstack.testing.pytest.container",
-    "localstack.testing.pytest.snapshot",
+    "localstack_snapshot.pytest.snapshot",
     "localstack.testing.pytest.filters",
     "localstack.testing.pytest.fixture_conflicts",
     "localstack.testing.pytest.detect_thread_leakage",
@@ -122,7 +132,49 @@ def pytest_collection_modifyitems(config, items):
 
 
 @pytest.fixture(scope="function")
-def snapshot(_snapshot_session: "SnapshotSession"):
+def snapshot(request, _snapshot_session: "SnapshotSession", aws_client):
+    account_id = aws_client.sts.get_caller_identity()["Account"]
+    region = aws_client.sts.meta.region_name
+
+    # Overwrite utility with our own => Will be refactored in the future
+    _snapshot_session.transform = TransformerUtility
+
+    _snapshot_session.add_transformer(RegexTransformer(account_id, "1" * 12), priority=2)
+    _snapshot_session.add_transformer(RegexTransformer(region, "<region>"), priority=2)
+
+    # TODO: temporary to migrate to new default transformers.
+    #   remove this after all exemptions are gone
+    exemptions = [
+        "tests/aws/services/acm",
+        "tests/aws/services/apigateway",
+        "tests/aws/services/cloudwatch",
+        "tests/aws/services/cloudformation",
+        "tests/aws/services/dynamodb",
+        "tests/aws/services/events",
+        "tests/aws/services/iam",
+        "tests/aws/services/kinesis",
+        "tests/aws/services/kms",
+        "tests/aws/services/lambda_",
+        "tests/aws/services/logs",
+        "tests/aws/services/route53",
+        "tests/aws/services/route53resolver",
+        "tests/aws/services/s3",
+        "tests/aws/services/secretsmanager",
+        "tests/aws/services/ses",
+        "tests/aws/services/sns",
+        "tests/aws/services/stepfunctions",
+        "tests/aws/services/sqs",
+        "tests/aws/services/transcribe",
+        "tests/aws/scenario/bookstore",
+        "tests/aws/scenario/note_taking",
+        "tests/aws/scenario/lambda_destination",
+        "tests/aws/scenario/loan_broker",
+    ]
+    if any([e in request.fspath.dirname for e in exemptions]):
+        _snapshot_session.add_transformer(SNAPSHOT_BASIC_TRANSFORMER, priority=2)
+    else:
+        _snapshot_session.add_transformer(SNAPSHOT_BASIC_TRANSFORMER_NEW, priority=2)
+
     return _snapshot_session
 
 
@@ -170,3 +222,11 @@ def secondary_aws_client(aws_client_factory):
     from localstack.testing.aws.util import secondary_testing_aws_client
 
     return secondary_testing_aws_client(aws_client_factory)
+
+
+@pytest.fixture(name="region", scope="session")
+def fixture_region(aws_client):
+    if is_aws() or is_api_enabled("sts"):
+        return aws_client.sts.meta.region_name
+    else:
+        return TEST_AWS_REGION_NAME
